0.12.4 (#163)
* 0.12.4

Update package dependencies and versions

This update includes several package dependency version upgrades in the package-lock.json file. This was necessary for ensuring that the dependencies of the project are up-to-date, potentially improving performance, security and compatibility with other packages or system features. Additionally, some packages were removed e.g., @typescript-eslint/* packages seems to be no longer needed and thus removed.