Update package dependencies versions (#168)
The versions of some dependencies in package-lock.json were incremented. This includes sentry packages, which have been upgraded to remove bugs and securities. Other changes include some packages having their revisions and other parameters updated. These changes are implemented to keep the application up-to-date as well as secure with the latest versions and patches.