0.12.10 (#169)
Update package version and always delete cookies on sign out

The package version was increased from 0.12.9 to 0.12.10 in package.json. In the sign-out server session script, the warning message was enriched by adding more context about what measures are taken when a user's session can't be destroyed. The code was also refactored to always delete the session cookies, regardless of the successful session destruction, for better safety and user experience.