0.13.3 (#173)
* 0.13.3

Remove unused components and update package version, use combined package for react.email

Deleted unused components CallToActionButton and EmailNavbar as they are no longer needed. Updated the package-lock with package version bumped from 0.13.1 to 0.13.2. This was done as part of project cleanup and optimizations.

* Allow dynamic selection of GitHub Actions runner

In `.github/workflows/build.yml`, configuration for jobs `lint`, `typecheck`, and `cypress` have been updated to support selection of GitHub Actions runner based on `env.RUNNER` environment variable. If `env.RUNNER` is not set, it will default to 'ubuntu-latest'. This change provides flexibility to run the workflows on alternate environments when necessary.

* Update GitHub Action to use vars instead of env

This commit fixes the ESLint job in the Github Actions workflow. It changes 'env' to 'vars' in the 'runs-on' field. This is done to adhere with the recent standardization changes in Github Actions that recommends using 'vars' instead of 'env'.

* Update node version and implement caching in Github Actions

The build script in Github actions was upgraded to make use of caching for npm dependencies. This change optimizes the build process by avoiding a fresh download of dependencies in case they are cached from previous runs. Along with this, the node.js setup version is also updated from 16 to 18 for maintaining compatibility with latest dependencies and to benefit from the improvements in the updated version. The container image in the cypress workflow was also updated

* Update GitHub Actions workflow for build efficiency

The changes removed the dependency caching and setup a specific Node version for the GitHub Actions workflow. This eliminated the caching of NPM dependencies and instead directly installs them, reducing complexity and potential cache-related issues in the CI/CD pipeline. Also,

* Add multilanguage support and theme selection to UI

This commit addresses two main changes:

1. Multilanguage support is added to the authentication provider, enabling firebase SDK to match the language settings specified within the application's context which was previously missing.

2. Small improvements are made to the UI, in the dropdown profile, allowing users to see their current selected theme with a check mark. The chosen theme is stored and retrieved appropriately. This provides a better user experience by visually confirming the selected user theme.