0.13.10 (#180)
* 0.13.10

Update package versions in package-lock.json

Updated multiple package versions in package-lock.json such as "@sentry/node", "@sentry/react", and "react-i18next".
These updates include both major and minor version changes. These changes occur as a consequence of regular maintenance and update procedures to keep the application up-to-date, secure, and stable with the latest package releases.

* Update color scheme of CheckoutRedirectButton

Changed the color scheme of the CheckoutRedirectButton in the 'subscriptions' component to improve the aesthetic and user experience. The color class was changed from text-foreground and bg-background to text-primary-foreground and bg-primary for better visibility. Additionally, the button variant was updated to 'outline' for a more modern and cohesive look across the application.