0.14.2 (#183)
* 0.14.2

Update email testing setup to use InBucket, replace sign-in with provider to redirect strategy (optional but set to true)

Upgraded the project to use InBucket for local email testing instead of Ethereal. This includes fetching the invite email in Cypress tests directly from the InBucket instance. Changes have also been made to use-sign-in-with-provider to handle both popup and redirect sign-in strategies. Besides, added an optional 'useRedirectStrategy' to the auth configuration. A number of packages have been updated including react and eslint. The package-lock.json has been updated to reflect these changes. The "inbucket:start" script has been added to start the InBucket service using Docker.

* Updated scripts, cypress tests and invitation mechanism

    Updated the start script for the inbucket service to be persistent and the email URL port has been updated to match the new settings. Refactored the invite member test to improve readability and grouped related tests. Updated the package script regarding the start of inbucket for better maintenance. Improved the organization invitation mechanism for better logging and consistent email delivery. Changes are aimed to improve test coverage, consistency, and clarity.

* Add error handling for ResizeObserver loop limit

This commit introduces a fix for the "ResizeObserver loop limit exceeded" exception. This error was being returned during automated tests and causing them to fail. The new error handling will avoid the test to stop due to this specific exception.