0.14.4 (#185)
* 0.14.4

Add functionality to delete and leave organization

Cypress tests have been added for delete and leave organization functionality. We also implemented the API to handle the 'leave organization' action and handled the scenario where the user trying to leave the organization is the owner. Updated the 'next' package and its plugins and subpackages alongside 'cypress'. We've created these functions to ensure the appropriate user access and removal from organizations under certain conditions.