0.15.2 (#190)
Update package version and refactor force sign out logic

The version of the 'next-firebase-saas-kit' package has been incremented to 0.15.2 for newer changes. In the 'with-app-props.ts' file, a 'forceSignOut' function has been created, which is now being used in the condition where user data cannot be fetched and in catch block, which helps in avoiding code duplication. This provides a cleaner and more maintainable code structure.