Add e2e tests for auth state and refactor indexedDB handling
Added Cypress end-to-end tests for logged in and logged out states in the file 'auth-state.cy.ts' to ensure user redirection to login page works as expected when there's no valid session. A console.log has also been added to the 'onAuthStateChanged' event for debug purposes.

Refactored how indexedDB is cleared in Cypress commands; it is now done within a 'cy.window()' to ensure it executes in the correct context.

The changes improve the reliability of automatic testing and ensures consistency in application behavior for different designs.