Update padding in AppHeader, increase package version & remove memoryPersistence (#191)
Updated the padding in 'AppHeaderWithMenu.tsx' component to improve UI appearance. The bottom-padding 'py-2' was added to balance the existing horizontal padding 'px-3'.

Incremented the package version in 'package.json' to '0.15.3' to reflect the new changes.

Removed inMemoryPersistence from FirebaseAuthProvider.tsx as it's not necessary. Since IndexedDB is fully supported in the browser environment, inMemoryPersistence is not needed.