0.16.0 (#193)
* 0.16.0

Remove obsolete and refactor several components to improve sidebars

The AppHeaderNoMenu component is removed since it no longer aligns with the style of the app. In the admin/organizations and admin organizations index pages, div elements are replaced with the PageBody element for consistency. The MobileNavigation component is enhanced by importing components that provide more functionality and user-friendliness. OrganizationsSelector is removed from the AppHeaderWithMenu.tsx file for cleaner design. Conversion from div elements to the Page element in RouteShellWithSidebar.tsx file improves readability and maintenance. For the RouteShell.tsx file, both title and description are soon to be required. For the SettingsPageContainer.tsx and SettingsTile.tsx files, they are refactored to use the Page component thus reducing redundancy. ProfileAvatar is updated to always display the user's name alongside their avatar in the sidebar. DashboardDemo no longer shows user greetings since it's unnecessary. The Subscription index page in the settings has its SettingsTile replaced by Section for better readability. The dropdown styling is slightly updated.

* Refactor components to allow React Node for title and description

Multiple files were updated to improve flexibility regarding title and description props. Now 'title' and 'description' props accept both string and React.ReactNode type. This allows another React component to be passed as a title or description, necessary for implementing localization through 'next-i18next'. Majority of the changes related to refactoring props in layout components such as 'RouteShellWithTopNavigation.tsx', 'RouteShellWithSidebar.tsx', etc

* Update admin labels and remove unused sidebar context

The labels in 'AdminHeader.tsx' and 'AdminRouteShell.tsx' were updated and the unused 'SidebarContext' in 'AdminProviders.tsx' was removed. Specifically, the "Back to Dashboard" label was changed to "Back to App" to make it more relevant. Additionally, the 'AdminProviders' component was simplified by removing the sidebar context, which was no longer required after recent changes.

* Update styling for ProfileAvatar component

Updated the `<Avatar>` component in the ProfileAvatar file by adding a class for margin-auto, width and height properties. This gives our avatar a fixed size to make it a more predictable and consistent component. The NavigationItem component styles were also refactored to enforce width and height to be full and padding was moved inside child context for proper alignment, leading to overall better appearance and ease of maintenance.

* Set default value for 'collapsed' in Sidebar component

Updated the Sidebar component to set a default value for the 'collapsed' parameter. This change ensures that if no value is explicitly set, the Sidebar won't be collapsed by default. This leads to an improved user experience, especially for first-time users who may be unfamiliar with our UI.

* Improve display and responsiveness of organization selector

Updated the OrganizationsSelector in light of improvements in accessibility and responsiveness. A role of "button" has been assigned to the div tag, which is helpful for screen readers. Changed element from span to div for better structure & responsiveness.  The placement of the logo and