Update package dependencies and increase project version (#199)
Updated several package dependencies in package.json and package-lock.json to use their newer versions. This includes common libraries such as firebase-admin, sentry, and stripe, among others. Alongside, this commit also includes incrementing the project version to 0.16.5. It's a routine dependency update to keep the project up-to-date with the latest libraries.