Refactor UpdateProfileForm to use UserSession (#229)
Refactored the UpdateProfileForm to replace Firebase User objects with UserSession objects, providing a better structure and more detailed user data. Also, redundant removal calls were optimized and moved to a separate function for cleaner code. The version was updated to 0.17.1 in package.json and package-lock.json. The UpdateOrganizationForm was also revised to use the component UploadLogoForm, making the code more organized and maintainable.